Upgrade GCloud App Engine instance to F2 to increase memory overhead
In F1 instance, programme was being shut down by supervisor for using too much
memory. I have not found a way of monitoring actual memory usage, so I am unsure
how much is being used, but there should at least not be memory leakage now, so
it is just the steady-state memory requirement of running a generative neural
net on the server.